requirements.txt
flask==3.0.3
gunicorn==22.0.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.2
alembic==1.13.2
python-dotenv==1.0.1
alembic==1.13.2
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.0.3
gunicorn==22.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.5
SQLAlchemy==2.0.36
typing_extensions==4.15.0
Werkzeug==3.1.3
cryptography==41.0.7
httpx[http2]==0.25.2
pillow==11.3.0
flask-cors==6.0.1
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0